// frontend/src/client/types.gen.ts
export type UpdatePassword = {
    current_password: string;
    new_password: string;
};
// ...
export type UserUpdateMe = {
    full_name?: (string | null);
    email?: (string | null);
};